Black sucks

+++ pySDC/playgrounds/deprecated/pmesh/playground_pmesh_comm.py
@@ -57,17 +57,29 @@ print(world_rank, time_rank, space_rank)
 t0 = time.perf_counter()
 
 if time_rank == 0:
-    pm = ParticleMesh(BoxSize=1.0, Nmesh=[nvars] * 2, dtype='f8', plan_method='measure', comm=space_comm)
-    u = pm.create(type='real')
-    u = u.apply(doublesine, kind='index', out=Ellipsis)
+    pm = ParticleMesh(
+        BoxSize=1.0,
+        Nmesh=[nvars] * 2,
+        dtype="f8",
+        plan_method="measure",
+        comm=space_comm,
+    )
+    u = pm.create(type="real")
+    u = u.apply(doublesine, kind="index", out=Ellipsis)
     time_comm.send(u.value, dest=1, tag=11)
 else:
-    pm = ParticleMesh(BoxSize=1.0, Nmesh=[nvars] * 2, dtype='f8', plan_method='measure', comm=space_comm)
+    pm = ParticleMesh(
+        BoxSize=1.0,
+        Nmesh=[nvars] * 2,
+        dtype="f8",
+        plan_method="measure",
+        comm=space_comm,
+    )
     tmp = time_comm.recv(source=0, tag=11)
-    u = pm.create(type='real', value=tmp)
+    u = pm.create(type="real", value=tmp)
 
 
 t1 = time.perf_counter()
 
-print(f'PMESH setup time: {t1 - t0:6.4f} sec.')
+print(f"PMESH setup time: {t1 - t0:6.4f} sec.")
 exit()
